kubectl wait: split condition functions into separate files
